% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/grover.R
\name{auth}
\alias{auth}
\alias{auth,Grover-method}
\title{auth}
\usage{
auth(grove)

\S4method{auth}{Grover}(grove)
}
\arguments{
\item{grove}{S4 object of class Grover}
}
\description{
Retrieve authentication key from a Grover object.
}
