% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{change_extent}
\alias{change_extent}
\title{A function to calculate change extent area and generate change rasters for
reporting.}
\usage{
change_extent(irast, rastkey, iregions, attribname, cloud = FALSE)
}
\arguments{
\item{irast}{Character file path to input veg density rasters that have been
through the veg classification process, i.e. those that are found in
`veg_class\`.}

\item{rastkey}{Character representation of unique string to match in intended
rasters. This aids in not selecting additional files generated by other
software (e.g. ArcMap). It will also be the output format of any rasters.}

\item{iregions}{Character file path to a shape file (including extension)
that defines reporting regions. The shape file should have an attribute
column that defines the overall reporting "region" plus "site", if
applicable, such as "NatPark_site1", "NatPark_site2" etc. The underscore
delineates region from site.}

\item{attribname}{Character string of the name of the attribute column that
contains the region information.}

\item{cloud}{Logical. Default is FALSE. If input imagery has  undergone
groveR cloud masking choose TRUE. Can add significant processesing time
on very large rasters.}
}
\value{
Rasters and calculated areas are exported to `extent_change\`.
}
\description{
\code{change_extent} creates change rasters between all consecutive  vegetation
     classification rasters and summary area stats.
}
\details{
This function is designed to do two things:
 \itemize{
  \item Firstly it creates change rasters between consecutive vegetation
      classification raster as produced by running \code{\link{veg_class}}. It
      summarises the densities into loss gain and stable categories in the
      output rasters.
  \item Secondly it generates area stats for regions and sites, defined by a
      user's shape file.
  }
}
\examples{
\dontrun{
change_extent(irast = "./veg_class", rastkey = ".img",
    iregions = "./vectors/regions.shp", attribname = "regions",
    cloud = FALSE)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
