% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{change_extent_plot}
\alias{change_extent_plot}
\title{A function to generate a stacked bar chart for mangrove change summaries.}
\usage{
change_extent_plot(icsv, cap)
}
\arguments{
\item{icsv}{Character representation of the name of the extent change
csv file including file path.}

\item{cap}{Character string of caption (not title) to supply to the plot. If
none required use empty quotes i.e. "".}
}
\value{
A PNG format graphic will be written to the `extent_change\`
    directory.
}
\description{
\code{change_extent_plot} creates stacked bar chart of vegetation change summaries.
}
\details{
Takes a file path to an extent change csv file as created by
    \code{link{change_extent}} located in `extent_change\` and produce
    a stacked bar chart representing the same.
}
\examples{
\dontrun{
change_extent_plot(icsv = "./extent_change/NatPark_2005-2009_extent_summaries.csv",
    cap = "RSSA")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
