% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-conversion.R
\name{conversionArgsPeakPick}
\alias{conversionArgsPeakPick}
\alias{conversionArgsNegativeMode}
\alias{conversionArgsPositiveMode}
\alias{conversionArgsMSlevel1}
\alias{conversionArgsMSlevel2}
\alias{conversionArgsMSlevel3}
\title{Conversion arguments}
\usage{
conversionArgsPeakPick()

conversionArgsNegativeMode()

conversionArgsPositiveMode()

conversionArgsMSlevel1()

conversionArgsMSlevel2()

conversionArgsMSlevel3()
}
\description{
Helper functions for msconvert arguments for raw data conversion.
}
\details{
These functions return character strings suitable for supplying to the \code{args}
argument of \code{convertFile} for passing to msconvert.
}
\examples{
conversionArgsPeakPick()
}
