% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conversionArgsIgnoreUnknownInstrumentError}
\alias{conversionArgsIgnoreUnknownInstrumentError}
\title{conversionArgsIgnoreUnknownInstrumentError}
\usage{
conversionArgsIgnoreUnknownInstrumentError()
}
\description{
msconvert args to ignore the unkown instrument error.
}
