% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conversionArgsNegativeMode}
\alias{conversionArgsNegativeMode}
\title{conversionArgsNegativeMode}
\usage{
conversionArgsNegativeMode()
}
\description{
msconvert args to convert only positive mode data.
}
