% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convertDirectory}
\alias{convertDirectory}
\title{convertDirectory}
\usage{
convertDirectory(outDir = ".", host, port, authKey, instrument,
  directory, args)
}
\description{
convert a directory of raw files using the grover API
}
