% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convertDirectorySplitModes}
\alias{convertDirectorySplitModes}
\title{convertDirectorySplitModes}
\usage{
convertDirectorySplitModes(outDir = ".", host, port, authKey, instrument,
  directory, args)
}
\description{
convert a directory of raw files, splitting positive and negative mode data using the grover API
}
