% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-conversion.R
\name{convertFile}
\alias{convertFile}
\title{convertFile}
\usage{
convertFile(authKey, instrument, directory, file, args = "", res)
}
\arguments{
\item{instrument}{instrument name}

\item{directory}{directory name}

\item{file}{file name}

\item{args}{arguments to pass to msconvert}

\item{grove}{S4 object of class Grover}

\item{outDir}{output directory path for converted files}
}
\description{
Convert a raw MS file using the grover API.
}
