% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-stats.R
\name{directoryFileInfo}
\alias{directoryFileInfo}
\alias{directoryFileInfo,GroverClient-method}
\title{directoryFileInfo}
\usage{
directoryFileInfo(grover_client, instrument, directory)

\S4method{directoryFileInfo}{GroverClient}(grover_client, instrument, directory)
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}

\item{instrument}{instrument name}

\item{directory}{directory name}
}
\description{
Return file information for all files in a given directory using 
the grover API.
}
