% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-check.R
\name{extant,GroverClient-method}
\alias{extant,GroverClient-method}
\title{extant}
\usage{
\S4method{extant}{GroverClient}(grover_client)
}
\arguments{
\item{grover_client}{S4 object of class GroverGlient}
}
\description{
Check grover API is still extistant.
}
