% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-stats.R
\name{fileInfo}
\alias{fileInfo}
\alias{fileInfo,GroverClient-method}
\title{fileInfo}
\usage{
fileInfo(grover_client, instrument, directory, file)

\S4method{fileInfo}{GroverClient}(grover_client, instrument, directory, file)
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}

\item{instrument}{instrument name}

\item{directory}{directory name}

\item{file}{file name}
}
\description{
Return file information for a given file using the grover API.
}
