% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{Grover-class}
\alias{Grover-class}
\title{Grover}
\description{
A simple S4 class containing grover API host information
}
\section{Slots}{

\describe{
\item{\code{host}}{host address}

\item{\code{port}}{port on which the API is hosted}

\item{\code{auth}}{authentication key}
}}

