% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grover.R
\name{grover}
\alias{grover}
\title{grover}
\usage{
grover(host, port, auth, repository = NULL)
}
\arguments{
\item{host}{host address}

\item{port}{port on which the API is hosted}

\item{auth}{authentication key}

\item{repository}{data repository directory path}
}
\description{
Create a GroverClient or GroverHost object containing API host information.
}
\examples{
grove <- grover('localhost',8000,'1234')
}
