% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/grover.R
\name{host}
\alias{host}
\alias{host<-}
\alias{port}
\alias{port<-}
\alias{auth}
\alias{auth<-}
\alias{host,GroverClient-method}
\alias{host<-,GroverClient-method}
\alias{port,GroverClient-method}
\alias{port<-,GroverClient-method}
\alias{auth,GroverClient-method}
\alias{auth<-,GroverClient-method}
\title{GroverClient get and set methods}
\usage{
host(grover_client)

host(grover_client) <- value

port(grover_client)

port(grover_client) <- value

auth(grover_client)

auth(grover_client) <- value

\S4method{host}{GroverClient}(grover_client)

\S4method{host}{GroverClient}(grover_client) <- value

\S4method{port}{GroverClient}(grover_client)

\S4method{port}{GroverClient}(grover_client) <- value

\S4method{auth}{GroverClient}(grover_client)

\S4method{auth}{GroverClient}(grover_client) <- value
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}

\item{value}{new host value to set}
}
\description{
Retrieve or set host information for a GroverClient objects.
}
