% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{GroverClient-class}
\alias{GroverClient-class}
\title{GroverClient class}
\description{
A simple S4 class containing grover API host information for client-side use.
}
\section{Slots}{

\describe{
\item{\code{host}}{host address}

\item{\code{port}}{port on which the API is hosted}

\item{\code{auth}}{authentication key}
}}

