% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/grover.R
\name{repository}
\alias{repository}
\alias{repository<-}
\alias{temp}
\alias{temp<-}
\alias{repository,GroverHost-method}
\alias{repository<-,GroverHost-method}
\alias{temp,GroverHost-method}
\alias{temp<-,GroverHost-method}
\title{GroverHost get and set methods}
\usage{
repository(grover_host)

repository(grover_host) <- value

temp(grover_host)

temp(grover_host) <- value

\S4method{repository}{GroverHost}(grover_host)

\S4method{repository}{GroverHost}(grover_host) <- value

\S4method{temp}{GroverHost}(grover_host)

\S4method{temp}{GroverHost}(grover_host) <- value
}
\arguments{
\item{grover_host}{S4 object of class GroverHost}

\item{value}{new host value to set}
}
\description{
Retrieve or set host information for GroverHost objects.
}
