% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{GroverHost-class}
\alias{GroverHost-class}
\title{GroverHost class}
\description{
A simple S4 class containing host information for configuring a grover API.
}
\details{
This class inherits from the GroverClient class so GroverClient accessor methods are applicable.
}
\section{Slots}{

\describe{
\item{\code{host}}{host address}

\item{\code{port}}{port on which the API is hosted}

\item{\code{auth}}{authentication key}

\item{\code{repository}}{directory path to the raw mass spectrometry data repository}

\item{\code{temp}}{optional directory path where converted files can temporarily be saved}
}}

