% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/grover.R
\name{host}
\alias{host}
\alias{host<-}
\alias{host,GroverClient-method}
\alias{host<-,GroverClient-method}
\title{host}
\usage{
host(grove)

host(grove) <- value

\S4method{host}{GroverClient}(grove)

\S4method{host}{GroverClient}(grove) <- value
}
\arguments{
\item{grove}{S4 object of class GroverClient}

\item{value}{new host value to set}
}
\description{
Retrieve host information from a Grover object.
}
