% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-stats.R
\name{instrumentFileInfo}
\alias{instrumentFileInfo}
\alias{instrumentFileInfo,GroverClient-method}
\title{instrumentFileInfo}
\usage{
instrumentFileInfo(grover_client, instrument)

\S4method{instrumentFileInfo}{GroverClient}(grover_client, instrument)
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}

\item{instrument}{instrument name}
}
\description{
Return file information for all files in a given instrument directory using the grover API.
}
