% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-list.R
\name{listInstruments}
\alias{listInstruments}
\alias{listDirectories}
\alias{listFiles}
\alias{listRawFiles}
\alias{listInstruments,GroverClient-method}
\alias{listDirectories,GroverClient-method}
\alias{listFiles,GroverClient-method}
\alias{listRawFiles,GroverClient-method}
\title{List raw data repository contents}
\usage{
listInstruments(grover_client)

listDirectories(grover_client, instrument)

listFiles(grover_client, instrument, directory)

listRawFiles(grover_client, instrument, directory)

\S4method{listInstruments}{GroverClient}(grover_client)

\S4method{listDirectories}{GroverClient}(grover_client, instrument)

\S4method{listFiles}{GroverClient}(grover_client, instrument, directory)

\S4method{listRawFiles}{GroverClient}(grover_client, instrument, directory)
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}

\item{instrument}{instrument name}

\item{directory}{directory name}
}
\description{
List available instruments, experiment directories or samples using a grover API.
}
