% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-conversion.R
\name{listDirectories}
\alias{listDirectories}
\title{listDirectories}
\usage{
listDirectories(authKey, instrument)
}
\arguments{
\item{instrument}{instrument name}

\item{grove}{S4 object of class Grover}
}
\description{
List all directories for a given instrument using the grover API.
}
