% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-list.R
\name{listFiles}
\alias{listFiles}
\alias{listRawFiles}
\alias{listFiles,GroverClient-method}
\title{listFiles}
\usage{
listFiles(grover_client, instrument, directory)

listRawFiles(grover_client, instrument, directory)

\S4method{listFiles}{GroverClient}(grover_client, instrument, directory)
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}

\item{instrument}{instrument name}

\item{directory}{directory name}
}
\description{
List all files present in a given directory using the grover API.
}
