% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-list.R
\name{listInstruments}
\alias{listInstruments}
\alias{listInstruments,GroverClient-method}
\title{listInstruments}
\usage{
listInstruments(grover_client)

\S4method{listInstruments}{GroverClient}(grover_client)
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}
}
\description{
List available instruments using the grover API.
}
