% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{listRawFiles}
\alias{listRawFiles}
\title{listRawFiles}
\usage{
listRawFiles(host, port, authKey, instrument, directory)
}
\description{
list all raw files present in a given directory using the grover API.
}
