% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make_example_data}
\alias{make_example_data}
\title{A function to set up example data for practice}
\usage{
make_example_data(p = ".")
}
\arguments{
\item{p}{Character file path for the processing folder (top level directory).
Defaults to current location, i.e. ".".}
}
\value{
Creates dummy user data.
}
\description{
\code{make_example_data}
}
\details{
This function extracts the internal example data, shipped with the
    package, and saves it into the processing folder structure as created by
    \code{link{make_folders}}.
}
\examples{
\dontrun{
make_example_data(p = ".")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
