% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make_folders}
\alias{make_folders}
\title{Functions to make directories and example data
A Function to create folders for user data}
\usage{
make_folders(p = ".")
}
\arguments{
\item{p}{Character file path for the processing folder (top level directory).
Defaults to current location, i.e. ".".}
}
\value{
Creates directories and sub-directories for user data.
}
\description{
\code{make_folders}
}
\details{
This function creates the directories and sub-directories required
    to store user data (or example data if working through the vignettes).
    These are:
\itemize{
    \item `raster_masks/`
    \item `raster_masks/cloud_masks/`
    \item `supplementary/`
    \item `vectors/`
    }
}
\examples{
\dontrun{
make_folders(p = ".")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
