% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make_index_rasters}
\alias{make_index_rasters}
\title{A function to create a folder and set up example index data}
\usage{
make_index_rasters(p = ".")
}
\arguments{
\item{p}{Character file path for the processing folder (top level directory).
Defaults to current location, i.e. ".".}
}
\value{
Creates a folder `index_rasters/` and dummy index data.
}
\description{
\code{make_index_rasters}
}
\details{
This function firstly creates a folder (`index_rasters/`) and then
    extracts the internal index example data, shipped with the package, and
    writes it to this location. This is handled separately to the other
    `make_` functions as this data and its location are only applicable to
    following examples in the package vignettes.
}
\examples{
\dontrun{
make_index_rasters(p = ".")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
