% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mask_product}
\alias{mask_product}
\title{A function to apply common masks to rasters}
\usage{
mask_product(irast, rastkey, choice, imask, maskkey)
}
\arguments{
\item{irast}{Character file path to directory veg density input rasters.}

\item{rastkey}{Character representation of unique string to match in intended
rasters. This aids in not selecting additional files generated by other
software (e.g. ArcMap).}

\item{choice}{Character vector of chosen year/s. If all years in a folder are
to be processed choice can be "all".}

\item{imask}{Character file path to directory of raster masks to apply.}

\item{maskkey}{Character representation of the file extension of the raster
masks.}
}
\value{
For each input raster a masked product will be written to file in a
    folder named `veg_dens_mskd/`. Masked values will be NA.
}
\description{
\code{mask_product} takes one or multiple input rasters and applies
    pre-generated masks to exclude areas or clean up rasters.
}
\details{
When producing raster products for an area there are often regions
    that occur at any time step that are not required for further analysis.
    Things like water bodies or non-target vegetation are examples. A folder
    full of different raster masks can be applied with this function.

    Raster masks are a raster indicating spatially where values of
    interest are but can be created with the intent to include or exclude.
    Masks must contain only 1's and NA values. Normally a value of 1 indicates
    the areas of interest however if a mask is designed to be used in an
    inverse manner then "INV" should be included in its file name. The
    function will search for these letters and reverse the mask accordingly.
}
\examples{
\dontrun{
mask_product(irast = "./veg_dens", rastkey = ".tif", choice = "all",
    imask = "Z:/DEC/projectXX/processing/raster_masks", maskkey = ".img")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
