% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msconvert.R
\name{msconvert}
\alias{msconvert}
\title{msconvert}
\usage{
msconvert(file, outPath = ".", args = "", verbose = T)
}
\arguments{
\item{file}{file to convert}

\item{outPath}{output directory path}

\item{args}{arguments to pass to msconvert}
}
\description{
Convert vender mass spectrometry file formats using msconvert
}
