% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/grover.R
\name{port}
\alias{port}
\alias{port<-}
\alias{port,GroverClient-method}
\alias{port<-,GroverClient-method}
\title{port}
\usage{
port(grove)

port(grove) <- value

\S4method{port}{GroverClient}(grove)

\S4method{port}{GroverClient}(grove) <- value
}
\arguments{
\item{grove}{S4 object of class GroverClient}

\item{value}{new port value to set}
}
\description{
Retrieve port information from a Grover object.
}
