% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{rawDirectory}
\alias{rawDirectory}
\title{rawDirectory}
\usage{
rawDirectory(grove, instrument, directory, outDir = ".")
}
\arguments{
\item{grove}{S4 object of class Grover}

\item{instrument}{instrument name}

\item{directory}{directory name}

\item{outDir}{output directory path for converted files}
}
\description{
Transfer a directory of raw files using the grover API.
}
