% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{rawFile}
\alias{rawFile}
\title{rawFile}
\usage{
rawFile(grove, instrument, directory, file, outDir = ".")
}
\arguments{
\item{grove}{S4 object of class Grover}

\item{instrument}{instrument name}

\item{directory}{directory name}

\item{file}{file name}

\item{outDir}{output directory path for converted files}
}
\description{
Transfer a raw MS file using the grover API.
}
