% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grover.R
\name{readGrover}
\alias{readGrover}
\title{Parse a grover configuration file}
\usage{
readGrover(path = "grover.yml")
}
\arguments{
\item{path}{file path to text file containing host information}
}
\description{
Read grover API host information from a text file.
}
\details{
The text file should contain 3 lines, the first the host address, 
the second the port on which it is hosted and lastly the authentication
 key needed
}
\examples{
## Grover client
grover_client <- system.file('grover_client.yml',
                              package = 'grover')
grover_client <- readGrover(grover_client)

## Grover host
grover_host <- system.file('grover_host.yml',
                              package = 'grover')
grover_host <- readGrover(grover_host)
}
