% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/grover.R
\name{repository}
\alias{repository}
\alias{repository<-}
\alias{repository,GroverHost-method}
\alias{repository<-,GroverHost-method}
\title{repository}
\usage{
repository(grove)

repository(grove) <- value

\S4method{repository}{GroverHost}(grove)

\S4method{repository}{GroverHost}(grove) <- value
}
\arguments{
\item{grove}{S4 object of class GroverHost}

\item{value}{new repository value to set}
}
\description{
Retrieve data repository directory path key 
from a GroverHost.
}
