% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-stats.R
\name{repositoryFileInfo}
\alias{repositoryFileInfo}
\alias{repositoryFileInfo,GroverClient-method}
\title{repositoryFileInfo}
\usage{
repositoryFileInfo(grover_client)

\S4method{repositoryFileInfo}{GroverClient}(grover_client)
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}
}
\description{
Return file information for all files in the data repository using 
the grover API.
}
