% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-info.R
\name{runInfo}
\alias{runInfo}
\alias{runInfo,GroverClient-method}
\title{runInfo}
\usage{
runInfo(grover_client, instrument, directory)

\S4method{runInfo}{GroverClient}(grover_client, instrument, directory)
}
\arguments{
\item{grover_client}{S4 object of class Grover}

\item{instrument}{instrument name}

\item{directory}{directory name}
}
\description{
Get sample meta information for a directory run using the 
grover API.
}
