% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-info.R
\name{sampleInfo}
\alias{sampleInfo}
\alias{sampleInfo,GroverClient-method}
\title{sampleInfo}
\usage{
sampleInfo(grover_client, instrument, directory, file)

\S4method{sampleInfo}{GroverClient}(grover_client, instrument, directory, file)
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}

\item{instrument}{instrument name}

\item{directory}{directory name}

\item{file}{file name}
}
\description{
get sample header information for a given raw file using the grover API.
}
