% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{show,Grover-method}
\alias{show,Grover-method}
\title{show-Grover}
\usage{
\S4method{show}{Grover}(object)
}
\arguments{
\item{object}{S4 object of class Grover}
}
\description{
Show method for class S4 Grover.
}
