% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-stats.R
\name{fileInfo}
\alias{fileInfo}
\alias{directoryFileInfo}
\alias{instrumentFileInfo}
\alias{repositoryFileInfo}
\alias{fileInfo,GroverClient-method}
\alias{directoryFileInfo,GroverClient-method}
\alias{instrumentFileInfo,GroverClient-method}
\alias{repositoryFileInfo,GroverClient-method}
\title{Retrieve file information}
\usage{
fileInfo(grover_client, instrument, directory, file)

directoryFileInfo(grover_client, instrument, directory)

instrumentFileInfo(grover_client, instrument)

repositoryFileInfo(grover_client)

\S4method{fileInfo}{GroverClient}(grover_client, instrument, directory, file)

\S4method{directoryFileInfo}{GroverClient}(grover_client, instrument, directory)

\S4method{instrumentFileInfo}{GroverClient}(grover_client, instrument)

\S4method{repositoryFileInfo}{GroverClient}(grover_client)
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}

\item{instrument}{instrument name}

\item{directory}{directory name}

\item{file}{file name}
}
\description{
Return file information for a given file using the grover API.
}
