% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-transfer.R
\name{transferDirectory}
\alias{transferDirectory}
\alias{transferDirectory,GroverClient-method}
\title{transferDirectory}
\usage{
transferDirectory(grover_client, instrument, directory, outDir = ".")

\S4method{transferDirectory}{GroverClient}(grover_client, instrument, directory, outDir = ".")
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}

\item{instrument}{instrument name}

\item{directory}{directory name}

\item{outDir}{output directory path for converted files}
}
\description{
Transfer a directory using the grover API.
}
