% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{trend_class_area}
\alias{trend_class_area}
\title{A function to calculate trend class classification area for reporting.}
\usage{
trend_class_area(irast, iregions, attribname)
}
\arguments{
\item{irast}{Character file path to the trend class raster for a distinct period
that has been written to the `trend_class\` directory. If multiple periods
(and rasters) exist, run this one at a time.}

\item{iregions}{Character file path to a shape file (including extension)
that defines reporting regions. The shape file should have an attribute
column that defines the overall reporting "region" plus "site", if
applicable, such as "NatPark_site1", "NatPark_site2" etc. The underscore
delineates region from site.}

\item{attribname}{Character string of the name of the attribute column that
contains the region information.}
}
\value{
Calculated areas are exported via csv file to `trend_class\`.
}
\description{
\code{trend_class_area} creates area stats for previously generated trend class
     rasters.
}
\details{
This function is designed to calculate the area of each trend class
    as defined by using the \code{\link{trend_class}} per reporting area or
    site.Reporting areas are defined by input shape file. Outputs results to
    csv.
}
\examples{
\dontrun{
trend_class_areas(irast = "trend_class/Test_National_Park_2014-2023_trendclass.tif",
    iregions = "vectors/regions.shp", attribname = "regions")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
