% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{veg_class}
\alias{veg_class}
\title{A function to perform vegetation classifications}
\usage{
veg_class(irast, ext = ".tif", classes = "supplementary/density_classes.csv")
}
\arguments{
\item{irast}{Character file path to input veg density rasters that have been
through the masking process, i.e. those that are found in
`veg_dens_mskd\` or `veg_dens_mskd_cld\`.}

\item{ext}{Character representation of the file extension of the input rasters.
Defaults to ".tif" as this is the preferred file type.}

\item{classes}{Character string of the name of the density classes
csv file including file path. Defaults to "./supplementary/density_classes.csv"
which works with the suggested project folder structure and workflow.}
}
\value{
All input rasters will be reclassified and be written to `veg_class/`.
}
\description{
\code{veg_class} performs a raster reclassification based on a user supplied
    set of meaningful bins.
}
\details{
This function is designed to take the continuous values, in say a
    raster of vegetation densities, and convert them to numeric classes based
    on some prior determination. The user must supply a csv of density
    classes with the following three column format so that each row forms a
    bin:
\itemize{
  \item lower - numerical low value for bin
  \item upper - numerical high value (<=) for bin
  \item reclass - numerical classification value
  }

   The numerical classification can take any single integer value except
   6. The integer 6 is reserved for any values that had a masked value of -99
   created by using "\code{link{cloud_mask}}.

   The function will perform some "probable" values classification in those
   areas that are masked with a -99 value. It will identify the same pixels
   in the previous year's raster and return those classified values plus 10.
   The reasoning here is that those pixels possibly would not have changed
   density but they can still be identified and removed (e.g. any value > 10)
   in downstream analysis.
}
\examples{
\dontrun{
veg_class(irast = "veg_dens_mskd_cld")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
