% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{veg_class_area}
\alias{veg_class_area}
\title{A function to calculate vegetation classification area for reporting.}
\usage{
veg_class_area(
  irast,
  iregions,
  attribname,
  areaname,
  ext = ".tif",
  probabilities = TRUE
)
}
\arguments{
\item{irast}{Character file path to input veg density rasters that have been
through the veg classification process, i.e. those that are found in
`veg_class_cloud_prob\` or `veg_class\`.}

\item{iregions}{Character file path to a shapefile (including extension)
that defines reporting regions. The shapefile should have an attribute
column that defines the overall reporting "region" plus "site", such as
"NatPark_site1", "NatPark_site2" etc. The underscore delineates region
from site and must be present.}

\item{attribname}{Character string of the name of the attribute column that
contains the region information.}

\item{areaname}{Character vector representing the geographical area that the
user is processing, e.g. marine park acronym. It will be used for inclusion
to the output csv name.}

\item{ext}{Character representation of the file extension of the input rasters.
Defaults to ".tif" as this is the preferred file type.}

\item{probabilities}{Boolean to indicate whether to calculate cloudy
probabilities. Default is TRUE.}
}
\value{
Calculated areas are exported via csv file to `extent_summaries\`.
}
\description{
\code{veg_class_area} calculates classified areas.
}
\details{
This function is designed to calculate the area of each class as
    defined by using the \code{\link{veg_class}} per reporting area and site.
    Reporting areas are defined by input shapefile. Output results to csv.

    Note that the attribute column of the input shapefile must contain a
    region name followed by a site name separated by an underscore.

    The function can also compute cloudy probabilities. This is when some
    mangroves have been masked in one year, the function looks to see what the
    classification was the previous year. If that year wasn't cloudy in that
    location then it most probably retains the same classification. This is
    then reported as a "cloud likely ..." class and this is done per pixel.

    If this behavior is not wanted, 'probabilities = FALSE', then any pixel
    identified as cloudy is simply reported as "cloud" in the area stats.
}
\examples{
\dontrun{
veg_class_area(irast = "veg_class", iregions = "vectors/regions.shp",
               attribname = "regions", areaname = "lgcsmp")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
