% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{veg_dens}
\alias{veg_dens}
\title{A function to turn vegetation index raster data into a vegetation density
raster for further analysis.}
\usage{
veg_dens(irast, rastkey, choice, index, ext, calibration)
}
\arguments{
\item{irast}{Character file path to input vegetation index rasters.}

\item{rastkey}{Character representation of unique string to match in intended
rasters. This aids in not selecting additional files generated by other
software (e.g. ArcMap).}

\item{choice}{Character vector of chosen year/s. If all years in a folder are
to be processed choice can be "all".}

\item{index}{Character representation of the index (should be part of the
input raster file name) which will be replaced in the new output file name.}

\item{ext}{Character representation of the desired file type output. Common
types include ".tif" and ".img".}

\item{calibration}{Character representation of the name of the calibration
csv file including file path.}
}
\value{
For each input raster a vegetation density raster of the same will be
    written to file in a folder named `veg_dens/`.
}
\description{
\code{veg_dens} takes an annual series of vegetation index rasters and
    converts them to a vegetation density (cover) product.
}
\details{
Density to vegetation index relationship must be established through
    prior analysis and the calibration file will provide the coefficients
    from this. The calibration file is a simple csv with a value in 5 columns
    named as follows:
\itemize{
  \item coef - the coefficient of the regression
  \item intercept -  the intercept of the regression
  \item multiple - a value to multiply by to bring output to a percentage
  \item lower - lower limit of acceptable vegetation density
  \item upper - upper limit of acceptable vegetation density
}
   All input rasters should have the same extents and cell size. Whilst it
   won't affect the performance of this function, downstream processing will
   fail. See package vignette for full explanation of processing work flow.
}
\examples{
\dontrun{
veg_dens(irast = "Z:/DEC/projectXX/mosaics", rastkey = ".ers",
    choice = c("1991", "2010"), index = "ndvi", ext = ".tif",
    calibration = "./supplementary/calibration.csv")
    }

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
