% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{veg_dens_class_plot}
\alias{veg_dens_class_plot}
\title{A function to generate a stacked bar chart for mangrove density classes.}
\usage{
veg_dens_class_plot(icsv, areaname, cap)
}
\arguments{
\item{icsv}{Character representation of the name of the extent summary
csv file including file path.}

\item{areaname}{Character string of desired monitoring area name for inclusion
to output png name.}

\item{cap}{Character string of caption (not title) to supply to the plot. If
none required use empty quotes i.e. "".}
}
\value{
A PNG format graphic will be written to the `extent_summaries\`
    directory.
}
\description{
\code{veg_dens_class_plot} creates stacked bar chart of vegetation densities.
}
\details{
Takes a file path to an extent summary csv file as created by
    \code{link{veg_class_area}} located in `extent_summaries\` and produce
    a stacked bar chart representing the same.
}
\examples{
\dontrun{
veg_dens_class_plot(icsv = "./extent_summaries/NatPark_2005-2009_extent_summaries.csv",
    areaname = "NatPark", cap = "RSSA")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
