% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-check.R
\name{version}
\alias{version}
\alias{version,GroverClient-method}
\title{Return the version of a grover API}
\usage{
version(grover_client)

\S4method{version}{GroverClient}(grover_client)
}
\arguments{
\item{grover_client}{S4 object of class GroverGlient}
}
\value{
The version number of the grover API.
}
\description{
Return the version number of a grover API.
}
