% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/grover.R
\name{auth}
\alias{auth}
\alias{auth<-}
\alias{auth,GroverClient-method}
\alias{auth<-,GroverClient-method}
\title{auth}
\usage{
auth(grove)

auth(grove) <- value

\S4method{auth}{GroverClient}(grove)

\S4method{auth}{GroverClient}(grove) <- value
}
\arguments{
\item{grove}{S4 object of class GroverClient}
}
\description{
Retrieve authentication key from a Grover object.
}
