% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conversionArgsPositiveMode}
\alias{conversionArgsPositiveMode}
\title{conversionArgsPositiveMode}
\usage{
conversionArgsPositiveMode()
}
\description{
msconvert args to convert only negative mode data.
}
