% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-conversion.R
\name{convertFile}
\alias{convertFile}
\alias{convertDirectory}
\alias{convertDirectorySplitModes}
\alias{convertFile,GroverClient-method}
\alias{convertDirectory,GroverClient-method}
\alias{convertDirectorySplitModes,GroverClient-method}
\title{Convert raw files to mzML format}
\usage{
convertFile(
  grover_client,
  instrument,
  directory,
  file,
  args = "",
  outDir = ".",
  zip = TRUE,
  overwrite = FALSE
)

convertDirectory(
  grover_client,
  instrument,
  directory,
  args = "",
  outDir = ".",
  zip = TRUE,
  overwrite = FALSE,
  exclude = character()
)

convertDirectorySplitModes(
  grover_client,
  instrument,
  directory,
  args = "",
  outDir = ".",
  zip = TRUE
)

\S4method{convertFile}{GroverClient}(
  grover_client,
  instrument,
  directory,
  file,
  args = NULL,
  outDir = ".",
  zip = TRUE,
  overwrite = FALSE
)

\S4method{convertDirectory}{GroverClient}(
  grover_client,
  instrument,
  directory,
  args = "",
  outDir = ".",
  zip = TRUE,
  overwrite = FALSE,
  exclude = character()
)

\S4method{convertDirectorySplitModes}{GroverClient}(
  grover_client,
  instrument,
  directory,
  args = "",
  outDir = ".",
  zip = TRUE
)
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}

\item{instrument}{instrument name}

\item{directory}{directory name}

\item{file}{file name}

\item{args}{arguments to pass to msconverteR::convert_files}

\item{outDir}{output directory path for converted files}

\item{zip}{zip converted file}

\item{overwrite}{overwrite local mzML files that already exist at the \code{outDir} path that have the same file name as the raw files to be converted}

\item{exclude}{A character vector of regular expression patterns for which raw files with matching patterns will not be converted.}
}
\value{
A vector of file paths to converted data files.
}
\description{
Convert a raw MS file using a grover API.
}
