% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-conversion.R
\name{convertDirectory}
\alias{convertDirectory}
\alias{convertDirectory,GroverClient-method}
\title{convertDirectory}
\usage{
convertDirectory(grover_client, instrument, directory, args = "", outDir = ".")

\S4method{convertDirectory}{GroverClient}(grover_client, instrument, directory, args = "", outDir = ".")
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}

\item{instrument}{instrument name}

\item{directory}{directory name}

\item{args}{arguments to pass to msconverteR::convert_files}

\item{outDir}{output directory path for converted files}
}
\description{
Convert a directory of raw files using the grover API.
}
