% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-conversion.R
\name{convertDirectory}
\alias{convertDirectory}
\title{convertDirectory}
\usage{
convertDirectory(grove, instrument, directory, args = "", outDir = ".")
}
\arguments{
\item{grove}{S4 object of class Grover}

\item{instrument}{instrument name}

\item{directory}{directory name}

\item{args}{arguments to pass to msconvert}

\item{outDir}{output directory path for converted files}
}
\description{
Convert a directory of raw files using the grover API.
}
