% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convertDirectorySplitModes}
\alias{convertDirectorySplitModes}
\title{convertDirectorySplitModes}
\usage{
convertDirectorySplitModes(grove, instrument, directory, args = "",
  outDir = ".")
}
\description{
convert a directory of raw files, splitting positive and negative mode data using the grover API
}
