% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-conversion.R
\name{convertFile}
\alias{convertFile}
\alias{convertFile,GroverClient-method}
\title{convertFile}
\usage{
convertFile(
  grover_client,
  instrument,
  directory,
  file,
  args = "",
  outDir = "."
)

\S4method{convertFile}{GroverClient}(
  grover_client,
  instrument,
  directory,
  file,
  args = "",
  outDir = "."
)
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}

\item{instrument}{instrument name}

\item{directory}{directory name}

\item{file}{file name}

\item{args}{arguments to pass to msconverteR::convert_files}

\item{outDir}{output directory path for converted files}
}
\description{
Convert a raw MS file using the grover API.
}
