% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convertFile}
\alias{convertFile}
\title{convertFile}
\usage{
convertFile(outDir = ".", host, port, authKey, instrument, directory,
  file, args = "")
}
\description{
convert a raw MS file using the grover API.
}
