% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-check.R
\name{extant}
\alias{extant}
\alias{extant,GroverClient-method}
\title{Check the existance of a grover API}
\usage{
extant(grover_client)

\S4method{extant}{GroverClient}(grover_client)
}
\arguments{
\item{grover_client}{S4 object of class GroverGlient}
}
\value{
\code{TRUE} if the API host is accessible and \code{FALSE} if not.
}
\description{
Check grover API is still extistant.
}
