% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grover.R
\name{grover}
\alias{grover}
\title{GroverClient and GroverHost class constructor}
\usage{
grover(host, port, auth, repository = NULL, temp = NULL)
}
\arguments{
\item{host}{host address}

\item{port}{port on which the API is hosted}

\item{auth}{authentication key}

\item{repository}{data repository directory path}

\item{temp}{optional directory path for grover hosts where temporary converted files can be saved}
}
\description{
Create a GroverClient or GroverHost class object containing 
API host information.
}
\examples{
## Grover client 
grover_client <- grover(host = "127.0.0.1",
                     port = 8000,
                     auth = "1234")

## Grover host
grover_host <- grover(host = "127.0.0.1",
                     port = 8000,
                     auth = "1234",
                     repository = system.file('repository',
                                               package = 'grover'))
}
