% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grover.R
\name{grover}
\alias{grover}
\title{grover}
\usage{
grover(host, port, auth, repository = NULL)
}
\arguments{
\item{host}{host address}

\item{port}{port on which the API is hosted}

\item{auth}{authentication key}

\item{repository}{data repository directory path}
}
\description{
Create a GroverClient or GroverHost object containing API host information.
}
\examples{
## Grover client 
grove_client <- grover(host = "127.0.0.1",
                     port = 8000,
                     auth = "1234")

## Grover host
grove_host <- grover(host = "127.0.0.1",
                     port = 8000,
                     auth = "1234",
                     repository = system.file('repository',
                                               package = 'grover'))
}
