% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{groverAPI}
\alias{groverAPI}
\title{Grover API}
\usage{
groverAPI(grove)
}
\arguments{
\item{grove}{S4 object of class Grover}
}
\description{
Run the grover REST API for file conversion.
}
\examples{
\dontrun{
groverAPI(readGrover())
}
}
