% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/grover.R
\name{repository}
\alias{repository}
\alias{repository<-}
\alias{repository,GroverHost-method}
\alias{repository<-,GroverHost-method}
\title{GroverHost get and set methods}
\usage{
repository(grover_host)

repository(grover_host) <- value

\S4method{repository}{GroverHost}(grover_host)

\S4method{repository}{GroverHost}(grover_host) <- value
}
\arguments{
\item{value}{new host value to set}

\item{grover_client}{S4 object of class GroverHost}
}
\description{
Retrieve or set host information for GroverHost objects.
}
