% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/grover.R
\docType{methods}
\name{host}
\alias{host}
\alias{host,Grover-method}
\title{host}
\usage{
host(grove)

\S4method{host}{Grover}(grove)
}
\arguments{
\item{grove}{S4 object of class Grover}
}
\description{
Retrieve host information from a Grover object.
}
