% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-info.R
\name{sampleInfo}
\alias{sampleInfo}
\alias{runInfo}
\alias{sampleInfo,GroverClient-method}
\alias{runInfo,GroverClient-method}
\title{Retrieve sample raw file header information}
\usage{
sampleInfo(grover_client, instrument, directory, file)

runInfo(grover_client, instrument, directory, exclude = character())

\S4method{sampleInfo}{GroverClient}(grover_client, instrument, directory, file)

\S4method{runInfo}{GroverClient}(grover_client, instrument, directory, exclude = character())
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}

\item{instrument}{instrument name}

\item{directory}{directory name}

\item{file}{file name}

\item{time_out}{maximum request time in seconds. This may need to be increased for larger directories.}
}
\description{
Get sample header information for a given raw file or directory using
a grover API.
}
