% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/client-list.R
\name{listDirectories}
\alias{listDirectories}
\alias{listDirectories,GroverClient-method}
\title{listDirectories}
\usage{
listDirectories(grover_client, instrument)

\S4method{listDirectories}{GroverClient}(grover_client, instrument)
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}

\item{instrument}{instrument name}
}
\description{
List all directories for a given instrument using 
the grover API.
}
