% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-conversion.R
\name{listInstruments}
\alias{listInstruments}
\title{listInstruments}
\usage{
listInstruments(grove)
}
\arguments{
\item{grove}{S4 object of class Grover}
}
\description{
List available instruments using the grover API.
}
