% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-list.R
\name{listRawFiles,GroverClient-method}
\alias{listRawFiles,GroverClient-method}
\title{listRawFiles}
\usage{
\S4method{listRawFiles}{GroverClient}(grover_client, instrument, directory)
}
\arguments{
\item{grover_client}{S4 object of class GroverClient}

\item{instrument}{instrument name}

\item{directory}{directory name}
}
\description{
List all raw files present in a given directory using 
the grover API.
}
