% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-conversion.R
\name{listRawFiles}
\alias{listRawFiles}
\title{listRawFiles}
\usage{
listRawFiles(grove, instrument, directory)
}
\arguments{
\item{grove}{S4 object of class Grover}

\item{instrument}{instrument name}

\item{directory}{directory name}
}
\description{
List all raw files present in a given directory using the grover API.
}
