% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-msconvert.R
\name{msconvert}
\alias{msconvert}
\title{msconvert}
\usage{
msconvert(file, outPath = ".", args = "", verbose = TRUE)
}
\arguments{
\item{file}{file to convert}

\item{outPath}{output directory path}

\item{args}{arguments to pass to msconvert}

\item{verbose}{should output be printed to the console}
}
\description{
Convert vender mass spectrometry file formats using msconvert
}
