% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/grover.R
\docType{methods}
\name{port}
\alias{port}
\alias{port,Grover-method}
\title{port}
\usage{
port(grove)

\S4method{port}{Grover}(grove)
}
\arguments{
\item{grove}{S4 object of class Grover}
}
\description{
Retrieve port information from a Grover object.
}
