% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grover.R
\name{readGrover}
\alias{readGrover}
\title{readGrover}
\usage{
readGrover(path = "~/grover.txt")
}
\arguments{
\item{path}{file path to text file containing host information}
}
\description{
Read grover API host information from a text file.
}
\details{
The text file should contain 3 lines, the first the host address, the second the port on which it is hosted and lastly the authentication key needed
}
\examples{
\dontrun{
grove <- readGrover('~/grover.txt')
}
}
