% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{runInfo}
\alias{runInfo}
\title{runInfo}
\usage{
runInfo(grove, instrument, directory)
}
\arguments{
\item{grove}{S4 object of class Grover}

\item{instrument}{instrument name}

\item{directory}{directory name}
}
\description{
get sample information for an instrument run using the grover API and create runinfo table.
}
