% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-info.R
\name{sampleScanFilters}
\alias{sampleScanFilters}
\title{sampleScanFilters}
\usage{
sampleScanFilters(authKey, instrument, directory, file)
}
\arguments{
\item{instrument}{instrument name}

\item{directory}{directory name}

\item{file}{file name}

\item{grove}{S4 object of class Grover}
}
\description{
get sample scan filters for a given file using the grover API.
}
