% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{trend_class}
\alias{trend_class}
\title{A function for calculating trend and trend class images}
\usage{
trend_class(irast, rastkey, end, period, classes)
}
\arguments{
\item{irast}{Character file path to input veg density rasters that have been
masked and found in `veg_dens_mskd_for_trends\`.}

\item{rastkey}{Character representation of unique string to match in intended
rasters. This aids in not selecting additional files generated by other
software (e.g. ArcMap).}

\item{end}{Numeric denoting final year of analysis.}

\item{period}{Numeric denoting intended length of trend analysis. Note that
periods are matched to class bins so ensure that appropriate period
length is entered otherwise results will be meaningless.}

\item{classes}{Character representation of the name of the trend class
csv file including file path.}
}
\value{
A raster of slope values and a raster of trend classes are written to
    `trend_class\` for the intended period of analysis.
}
\description{
\code{trend_class} creates trend and trend class rasters from user specified
     time periods.
}
\details{
Given an end year and a period length this function stacks the
    appropriate annual rasters and outputs a linear trend raster and a trend
    class raster based on a user supplied classification in csv format. The
    csv must have the following three column format so that each row forms a
    bin:
\itemize{
  \item lower - numerical low value for bin
  \item upper - numerical high value (<=) for bin
  \item reclass - numerical classification value
  }
    This classification is based on binning the slope values from the
    regression and is therefore sensitive to the period length. Ensure that
    the classification matches the intended period of analysis.

    Note to accommodate any missing annual rasters, due to for example
    availability of suitable imagery, the function will create "dummy"
    rasters to pad the stack accordingly. These "dummy" rasters ensure the
    time variable is handled correctly in the regression. These "dummies" are
    deleted as the function cleans up after itself.
}
\examples{
\dontrun{
trend_class(irast = "./veg_dens_mskd_for_trends", rastkey = ".tif",
    end = 2009, period = 10, classes = "./supplementary/trend_classes.csv")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
