% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{veg_class_area}
\alias{veg_class_area}
\title{A function to calculate vegetation classification area for reporting.}
\usage{
veg_class_area(irast, rastkey, iregions, attribname, areaname)
}
\arguments{
\item{irast}{Character file path to input veg density rasters that have been
through the veg classification process, i.e. those that are found in
`veg_class_cloud_prob\` or `veg_class\`.}

\item{rastkey}{Character representation of unique string to match in intended
rasters. This aids in not selecting additional files generated by other
software (e.g. ArcMap).}

\item{iregions}{Character file path to a shape file (including extension)
that defines reporting regions. The shape file should have an attribute
column that defines the overall reporting "region" plus "site", if
applicable, such as "NatPark_site1", "NatPark_site2" etc. The underscore
delineates region from site.}

\item{attribname}{Character string of the name of the attribute column that
contains the region information.}

\item{areaname}{Character string of desired monitoring area name for inclusion
to output csv name.}
}
\value{
Calculated areas are exported via csv file to `extent_summaries\`.
}
\description{
\code{veg_class_area} calculates classified areas.
}
\details{
This function is designed to calculate the area of each class as
    defined by using the \code{\link{veg_class}} per reporting area or site.
    Reporting areas are defined by input shape file. Output results to csv.
}
\examples{
\dontrun{
veg_class_area(irast = "./veg_class_cloud_prob", rastkey = ".tif",
    iregions = "./vectors/regions.shp", attribname = "regions",
    areaname = "NKMP_all")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
