% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{veg_dens}
\alias{veg_dens}
\title{A function to turn vegetation index raster data into a vegetation density
raster for further analysis.}
\usage{
veg_dens(
  irast,
  areaname,
  ext = ".tif",
  calibration = "supplementary/calibration.csv"
)
}
\arguments{
\item{irast}{Character file path to input vegetation index rasters.}

\item{areaname}{Character vector representing the geographical area that the
user is processing, e.g. marine park name. This will become part of the
named interim data sets and should be brief and contain no spaces or
underscores. Acronyms are good.}

\item{ext}{Character representation of the input file type. Defaults to ".tif"
as this is the preferred file type.}

\item{calibration}{Character representation of the name of the calibration
csv file including file path. Defaults to "./supplementary/calibration.csv"
which works with the suggested project folder structure and workflow.}
}
\value{
For each input raster a vegetation density raster of the same will be
    written to file in a folder named `veg_dens/`.
}
\description{
\code{veg_dens} takes an annual series of vegetation index raster mosaics and
    converts them to a vegetation density (cover) product.
}
\details{
Density to vegetation index relationship must be established through
    prior analysis and the calibration file will provide the coefficients
    from this. The calibration file is a simple csv with a value in 5 columns
    named as follows:
\itemize{
  \item coef - the coefficient of the regression
  \item intercept -  the intercept of the regression
  \item multiple - a value to multiply by to bring output to a percentage
  \item lower - lower limit of acceptable vegetation density
  \item upper - upper limit of acceptable vegetation density
}
     All input raster mosaics should have the same extents and cell size. Whilst it
     won't affect the performance of this function, downstream processing will
     fail. See package vignette for full explanation of processing work flow.

     Input raster mosaics can follow any naming convention that makes sense to
     the project but must contain a 4 digit representation of the year and
     contain NO other numerals. An example might be "lgcsmp_ndvi_2023.tif".
}
\examples{
\dontrun{
veg_dens(irast = "Z:/DEC/projectXX/mosaics", areaname = "lgcsmp")
    }

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
