% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-check.R
\name{version}
\alias{version}
\title{Return the version of a grover API}
\usage{
version(grover_client)
}
\value{
the version number of the grover API
}
\description{
Return the version number of a grover API.
}
